chore: move stats to it's own file - correct spelling of stat - if statement refactor
